extract ibft collectors to a dedicated file
