+++ ibft/metrics.go
@@ -1,0 +1,34 @@
+package ibft
+
+import (
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
+	"log"
+)
+
+var (
+	metricsCurrentSequence = promauto.NewGaugeVec(prometheus.GaugeOpts{
+		Name: "ssv:validator:ibft_current_sequence",
+		Help: "The highest decided sequence number",
+	}, []string{"lambda", "pubKey"})
+	metricsIBFTStage = promauto.NewGaugeVec(prometheus.GaugeOpts{
+		Name: "ssv:validator:ibft_stage",
+		Help: "IBFTs stage",
+	}, []string{"lambda", "pubKey"})
+	metricsIBFTRound = promauto.NewGaugeVec(prometheus.GaugeOpts{
+		Name: "ssv:validator:ibft_round",
+		Help: "IBFTs round",
+	}, []string{"lambda", "pubKey"})
+)
+
+func init() {
+	if err := prometheus.Register(metricsCurrentSequence); err != nil {
+		log.Println("could not register prometheus collector")
+	}
+	if err := prometheus.Register(metricsIBFTStage); err != nil {
+		log.Println("could not register prometheus collector")
+	}
+	if err := prometheus.Register(metricsIBFTRound); err != nil {
+		log.Println("could not register prometheus collector")
+	}
+}
